requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
matplotlib>=3.7.0
scipy>=1.11.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
xlrd>=2.0.1
requests>=2.31.0
networkx>=3.0
adjustText>=0.8
kaleido>=0.2.1
python-docx>=1.1.0
nbformat>=5.9.0
pytest>=7.4.0
pytest-mock>=3.11.0
